backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend for Video Dubbing"
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    # ── Server ──────────────────────────────────────────────
    "fastapi",
    "uvicorn[standard]",
    "python-multipart",

    # ── ML core ─────────────────────────────────────────────
    "torch==2.3.1",
    "torchaudio==2.3.1",

    # ── Stage 2: Transcription ───────────────────────────────
    "whisperx",

    # ── Stage 3: Translation ─────────────────────────────────
    "requests",           # for Gemini API calls
    "python-dotenv",      # for environment variables

    # ── Stage 4: TTS (Sarvam AI API) ─────────────────────────
    "indic-nlp-library",  # text normalization for Indian scripts
    "soundfile",          # write .wav clips

    # ── Shared utils ─────────────────────────────────────────
    "pydub",
    "ffmpeg-python",
    "numpy<2",
    "setuptools<81",
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cu121" },
]
torchaudio = [
  { index = "pytorch-cu121" },
]

[[tool.uv.index]]
name = "pytorch-cu121"
url = "https://download.pytorch.org/whl/cu121"
explicit = true
